// client/src/app/models/eRole.ts
export enum ERole {
  ROLE_USER = 'ROLE_USER',
  ROLE_SECRETAIRE = 'ROLE_SECRETAIRE',
  ROLE_BUREAU = 'ROLE_BUREAU',
  ROLE_MEMBRECA = 'ROLE_MEMBRECA',
  ROLE_ADMIN = 'ROLE_ADMIN',
  ROLE_COMPTABLE = 'ROLE_COMPTABLE',
  ROLE_ENCADRANT = 'ROLE_ENCADRANT',
  ROLE_COMMUNICATION_GLOBAL = 'ROLE_COMMUNICATION_GLOBAL',
  ROLE_RESPONSABLE_BOUTIQUE = 'ROLE_RESPONSABLE_BOUTIQUE'
}